Move interface and Calculator to EasyInterface
