+++ App/PyImports/DisplayModels/FileStructureModel.py
@@ -3,7 +3,7 @@ import logging
 from PySide2.QtCore import Qt, QObject, Signal
 from PySide2.QtGui import QStandardItemModel, QStandardItem
 
-from PyImports.Models.BaseModel import BaseModel
+from PyImports.DisplayModels.BaseModel import BaseModel
 
 class FileStructureModel(BaseModel):
     def __init__(self, parent=None):
