+++ App/PyImports/DisplayModels/StatusModel.py
@@ -1,9 +1,9 @@
-from PySide2.QtCore import Qt, QObject
+from PySide2.QtCore import Qt
 from PySide2.QtGui import QStandardItem, QStandardItemModel
 
-import PyImports.Helpers as Helpers
-from PyImports.StatusObjects import StatusItem, StatusList
-from PyImports.Models.BaseModel import BaseModel
+import PyImports.EasyInterface.Utils.Helpers as Helpers
+from PyImports.EasyInterface import StatusItem, StatusList
+from PyImports.DisplayModels.BaseModel import BaseModel
 
 class StatusModel(BaseModel):
     def __init__(self, parent=None):
