+++ App/PyImports/EasyInterface/requirements.txt
@@ -1,0 +1,2 @@
+cryspy>=0.2.0
+dictdiffer
