+++ Tests/pytest/App/PyImports/Models/test_MeasuredDataModel.py
@@ -5,7 +5,7 @@ from PySide2.QtGui import QStandardItemModel
 
 from PyImports.Calculators.CryspyCalculator import CryspyCalculator
 
-import PyImports.Models.MeasuredDataModel as Model
+import PyImports.DisplayModels.MeasuredDataModel as Model
 
 TEST_FILE = "file:Tests/Data/main.cif"
